+++ shared/core/src/lib/utils/es-toolkit.ts
@@ -2,6 +2,7 @@ export {
   assignIn,
   dropRight,
   endsWith,
+  has,
   isEmpty,
   isFunction,
   join,