+++ shared/framework-kernel/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './utilities';
+export * from './stores';
 export * from './main';