+++ shared/framework-kernel/src/lib/utilities/index.ts
@@ -1,1 +1,1 @@
-export * from './RouterUtilities';
+export * from './router-utilities';