+++ shared/apis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @herodotus-cloud/apis
 
+## 1.6.3
+
+### Patch Changes
+
+- 迁移 useRouterStore 代码
+- Updated dependencies
+  - @herodotus-cloud/core@1.6.3
+
 ## 1.6.2
 
 ### Patch Changes