+++ shared/form-engine/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @herodotus-cloud/form-engine
 
+## 1.6.3
+
+### Patch Changes
+
+- 迁移 useRouterStore 代码
+- Updated dependencies
+  - @herodotus-cloud/form-apis@1.6.3
+
 ## 1.6.2
 
 ### Patch Changes